refactor of umd,commonjs imports, refactor of hhCalendar for local inclusion
