+++ components/04-reference-pages/_page-preview.hbs
@@ -21,7 +21,7 @@
    {{{ yield }}}
    <!-- DO NOT EDIT BELOW THIS POINT -->
    <script src="{{ path '/college-dls/js/bundle/college-dls-umd.js' }}" type="text/javascript"></script>
-   <script src="{{ path '/college-dls/js/back-to-top.js' }}" type="text/javascript"></script>
+   {{!-- <script src="{{ path '/college-dls/js/back-to-top.js' }}" type="text/javascript"></script> --}}
    <!-- DO NOT EDIT -->
 </body>
