+++ package.json
@@ -34,6 +34,7 @@
     "select2": "^4.1.0-rc.0"
   },
   "devDependencies": {
+    "@rollup/plugin-inject": "^5.0.5",
     "@rollup/plugin-node-resolve": "^15.0.2",
     "autoprefixer": "^10.4.13",
     "cssnano": "^5.1.14",
@@ -48,6 +49,8 @@
     "gulp-sass-glob": "^1.1.0",
     "gulp-uglify": "^3.0.2",
     "postcss": "^8.4.21",
+    "rollup-plugin-commonjs": "^10.1.0",
+    "rollup-plugin-node-resolve": "^5.2.0",
     "rollup-plugin-replace": "^2.2.0",
     "sass": "^1.57.1"
   },
